+++ app/controllers/application_controller.rb
@@ -8,4 +8,13 @@ class ApplicationController < ActionController::Base
   end
 
 
+  def submit_contact
+    if params[:name].present? and params[:email].present? and params[:body].present?
+      # UserMailer.contact(params[:name], params[:email], params[:body]).deliver_later if Rails.env.production?
+      redirect_to contact_path, notice: "Thanks for getting in touch!"
+    else
+      render :contact
+    end
+  end
+
 end